modified subset widget to not break our code
